feat: Implement Phase 2 - Workload modules (create_delete, rename, hardlink, mixed)
WORKLOADS IMPLEMENTED:
1. create_delete.c     - Original workload extracted (30% create, 10% delete)
2. rename.c            - Rename-heavy (30% create, 30% delete, 40% rename)
3. hardlink.c          - Hard link stress (30% create, 40% link, 30% unlink)
4. mixed.c             - All operations (25% create, 20% delete, 25% rename, 20% link, 10% symlink)

FEATURES:
- Each workload implements workload_ops_t interface
- Custom reader/writer threads per workload
- Workload-specific statistics tracking
- Shared reader functions (create_delete_reader, hardlink_reader)
- All integrated into workload registry

BUILD STATUS:
✅ All workloads compile successfully
✅ Registry links all 4 workloads
⚠️  BPF injectors have compilation issues (TODO: fix PT_REGS_RC macro)

CODE STATS:
- create_delete.c: 220 lines
- rename.c: 175 lines
- hardlink.c: 290 lines
- mixed.c: 223 lines
Total: 908 lines of workload code

NEXT:
- Fix BPF compilation (PT_REGS_RC issues with __TARGET_ARCH_x86)
- Implement Phase 3: Generic controller
- Implement Phase 4: Unified test runner

BACKWARD COMPATIBLE: simple_chaos_test.c still works unchanged

+++ chaos/injectors/rename_window.bpf.c
@@ -13,9 +13,11 @@
  * Based on research: docs/design/BTRFS-LOST-UPDATE-ANALYSIS.md
  */
 
+#include <linux/types.h>
 #include <linux/bpf.h>
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_tracing.h>
+#include <bpf/bpf_core_read.h>
 
 char LICENSE[] SEC("license") = "GPL";
 
@@ -99,7 +101,7 @@ static __always_inline void inject_delay(void) {
 
 SEC("kretprobe/__btrfs_unlink_inode")
 int hook_btrfs_unlink_exit(struct pt_regs *ctx) {
-    int ret = PT_REGS_RC(ctx);
+    long ret = PT_REGS_RC(ctx);
     
     // Update call count
     __u32 stat_key = STAT_BTRFS_UNLINK_CALLS;
@@ -162,7 +164,7 @@ int hook_vfs_rename_entry(struct pt_regs *ctx) {
 
 SEC("kretprobe/do_unlinkat")
 int hook_unlinkat_exit(struct pt_regs *ctx) {
-    int ret = PT_REGS_RC(ctx);
+    long ret = PT_REGS_RC(ctx);
     
     // Update total calls
     __u32 stat_key = STAT_TOTAL_CALLS;
